Separated code for filters/effects page
